refactoring out things from the mega models.py
