+++ app/server/api/geolocation_api.py
@@ -2,7 +2,7 @@ from flask import Blueprint, request, make_response, jsonify
 from flask.views import MethodView
 
 from server import db
-from server.models import models
+from server.models.user import User
 from server.utils.auth import requires_auth
 
 geolocation_blueprint = Blueprint('geolocation', __name__)
@@ -17,7 +17,7 @@ class AddAddressLatLng(MethodView):
         lat = post_data['lat']
         lng = post_data['lng']
 
-        user = models.User.query.get(user_id)
+        user = User.query.get(user_id)
 
         user.lat = lat
         user.lng = lng
