+++ app/server/api/recognised_face_api.py
@@ -2,7 +2,7 @@ from flask import Blueprint, request, make_response, jsonify
 from flask.views import MethodView
 
 from server import db
-from server.models.models import UploadedImage
+from server.models.upload import UploadedImage
 from server.utils.auth import requires_auth
 
 from sqlalchemy.orm.attributes import flag_modified
