+++ app/server/me_api/credit_transfer.py
@@ -12,15 +12,12 @@ from server.exceptions import (
     AccountNotApprovedError,
     InsufficientBalanceError
 )
-from server.models.models import (
-    User,
-    TransferAccount,
-    TransferCard,
-    CreditTransfer,
-    paginate_query
-)
+from server.models.user import User
+from server.models.transfer import TransferAccount, TransferCard, CreditTransfer
+from server.models.models import paginate_query
 from server.schemas import me_credit_transfers_schema, me_credit_transfer_schema
-from server.utils.auth import requires_auth, AccessControl, show_all
+from server.utils.auth import requires_auth, show_all
+from server.utils.access_control import AccessControl
 from server.utils.credit_transfers import (
     check_for_any_valid_hash,
     find_user_with_transfer_account_from_identifiers,
