+++ app/server/models/utils.py
@@ -1,4 +1,6 @@
+from flask import g
 import datetime
+from sqlalchemy.ext.declarative import declared_attr
 
 from server import db
 
@@ -10,6 +12,22 @@ def get_authorising_user_id():
     else:
         return None
 
+user_transfer_account_association_table = db.Table(
+    'user_transfer_account_association_table',
+    db.Model.metadata,
+    db.Column('user_id', db.Integer, db.ForeignKey('user.id')),
+    db.Column('transfer_account_id', db.Integer, db.ForeignKey('transfer_account.id'))
+)
+
+organisation_association_table = db.Table(
+    'organisation_association_table',
+    db.Model.metadata,
+    db.Column('organisation_id', db.Integer, db.ForeignKey('organisation.id')),
+    db.Column('user_id', db.Integer, db.ForeignKey('user.id')),
+    db.Column('transfer_account_id', db.Integer, db.ForeignKey('transfer_account.id')),
+    db.Column('credit_transfer_id', db.Integer, db.ForeignKey('credit_transfer.id')),
+)
+
 class ModelBase(db.Model):
     __abstract__ = True
 
@@ -17,3 +35,36 @@ class ModelBase(db.Model):
     authorising_user_id = db.Column(db.Integer, default=get_authorising_user_id)
     created = db.Column(db.DateTime, default=datetime.datetime.utcnow)
     updated = db.Column(db.DateTime, default=datetime.datetime.utcnow, onupdate=datetime.datetime.utcnow)
+
+class OneOrgBase(object):
+    """
+    Mixin that automatically associates object(s) to organisation(s) many-to-one.
+
+    Forces all database queries on associated objects to provide an organisation ID or specify show_all=True flag
+    """
+
+    @declared_attr
+    def organisation_id(cls):
+        return db.Column('organisation_id', db.ForeignKey('organisation.id'))
+
+
+class ManyOrgBase(object):
+    """
+    Mixin that automatically associates object(s) to organisation(s) many-to-many.
+
+    Forces all database queries on associated objects to provide an organisation ID or specify show_all=True flag
+    """
+
+    @declared_attr
+    def organisations(cls):
+
+        # pluralisation
+        name = cls.__tablename__.lower()
+        plural = name + 's'
+        if name[-1] in ['s', 'sh', 'ch', 'x']:
+            # exceptions to this rule...
+            plural = name + 'es'
+
+        return db.relationship("Organisation",
+                               secondary=organisation_association_table,
+                               back_populates=plural)
