+++ app/server/utils/phone.py
@@ -10,7 +10,6 @@ def make_sms_respone(message):
     resp.message(message)
     return str(resp)
 
-
 def proccess_phone_number(phone_number, region=None, ignore_region=False):
     """
     Parse any given phone number.
